Update: add permission when update item

+++ core/views/item_view.py
@@ -1,5 +1,6 @@
 from rest_framework import views, status, permissions
 from rest_framework.response import Response
+from utils.enum import ROLE
 
 from core.models.item import Item
 from core.serializers.item_serializer import ItemSerializer
@@ -16,14 +17,17 @@ class ItemAPIView(views.APIView):
         return Response(prepare_success_response(serializer.data), status=status.HTTP_200_OK)
 
     def post(self, request):
-        validate_error = validate_item_service(request.data)
-        if validate_error is not None:
-            return Response(prepare_error_response(validate_error), status=status.HTTP_400_BAD_REQUEST)
-        serializer = ItemSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save(proprietor=self.request.user.shop_owner)
-            return Response(prepare_create_success_response(serializer.data), status=status.HTTP_201_CREATED)
-        return Response(prepare_error_response(serializer.errors), status=status.HTTP_400_BAD_REQUEST)
+        if request.user.role == ROLE.ADMIN or request.user.role == ROLE.MANAGER or request.user.role == ROLE.SHOPKEEPER:
+            validate_error = validate_item_service(request.data)
+            if validate_error is not None:
+                return Response(prepare_error_response(validate_error), status=status.HTTP_400_BAD_REQUEST)
+            serializer = ItemSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save(proprietor=self.request.user.shop_owner)
+                return Response(prepare_create_success_response(serializer.data), status=status.HTTP_201_CREATED)
+            return Response(prepare_error_response(serializer.errors), status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response(prepare_error_response('You have no permission'), status=status.HTTP_400_BAD_REQUEST)
 
 
 class ItemUpdateDetailDeleteAPIView(views.APIView):
@@ -36,18 +40,21 @@ class ItemUpdateDetailDeleteAPIView(views.APIView):
             return None
 
     def put(self, request, pk):
-        validate_error = validate_item_service(request.data)
-        if validate_error is not None:
-            return Response(prepare_error_response(validate_error), status=status.HTTP_400_BAD_REQUEST)
-        item = self.get_object(pk)
-        if item is not None:
-            serializer = ItemSerializer(item, data=request.data)
-            if serializer.is_valid():
-                serializer.save()
-                return Response(prepare_create_success_response(serializer.data), status=status.HTTP_201_CREATED)
-            return Response(prepare_error_response(serializer.errors), status=status.HTTP_400_BAD_REQUEST)
+        if request.user.role == ROLE.ADMIN or request.user.role == ROLE.MANAGER or request.user.role == ROLE.SHOPKEEPER:
+            validate_error = validate_item_service(request.data)
+            if validate_error is not None:
+                return Response(prepare_error_response(validate_error), status=status.HTTP_400_BAD_REQUEST)
+            item = self.get_object(pk)
+            if item is not None:
+                serializer = ItemSerializer(item, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return Response(prepare_create_success_response(serializer.data), status=status.HTTP_201_CREATED)
+                return Response(prepare_error_response(serializer.errors), status=status.HTTP_400_BAD_REQUEST)
+            else:
+                return Response(prepare_error_response("No data found for this ID"), status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(prepare_error_response("No data found for this ID"), status=status.HTTP_400_BAD_REQUEST)
+            return Response(prepare_error_response('You have no permission'), status=status.HTTP_400_BAD_REQUEST)
 
     def get(self, request, pk):
         item = self.get_object(pk)
@@ -57,9 +64,11 @@ class ItemUpdateDetailDeleteAPIView(views.APIView):
         return Response(prepare_error_response("Content Not found"), status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
-        item = self.get_object(pk)
-        if item is not None:
-            item.delete()
-            return Response(prepare_success_response("Data deleted successfully"), status=status.HTTP_200_OK)
-        else:
-            return Response(prepare_error_response("Content Not found"), status=status.HTTP_400_BAD_REQUEST)
+        if request.user.role == ROLE.ADMIN or request.user.role == ROLE.MANAGER or request.user.role == ROLE.SHOPKEEPER:
+            item = self.get_object(pk)
+            if item is not None:
+                item.delete()
+                return Response(prepare_success_response("Data deleted successfully"), status=status.HTTP_200_OK)
+            else:
+                return Response(prepare_error_response("Content Not found"), status=status.HTTP_400_BAD_REQUEST)
+        return Response(prepare_error_response('You have no permission'), status=status.HTTP_400_BAD_REQUEST)
